# src/scripts/api-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Customers Lineup Api",
    "version": "v1"
  },
  "paths": {
...
    "/api/business": {
      "post": {
        "tags": [
          "Business"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CLup.API.Contracts.Businesses.CreateBusiness.CreateBusinessRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CLup.API.Contracts.Businesses.CreateBusiness.CreateBusinessRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CLup.API.Contracts.Businesses.CreateBusiness.CreateBusinessRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Business"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CLup.API.Contracts.Businesses.UpdateBusiness.UpdateBusinessRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CLup.API.Contracts.Businesses.UpdateBusiness.UpdateBusinessRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CLup.API.Contracts.Businesses.UpdateBusiness.UpdateBusinessRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/message/user": {
      "post": {
        "tags": [
          "Message"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CLup.API.Contracts.Messages.SendUserMessage.SendUserMessageRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CLup.API.Contracts.Messages.SendUserMessage.SendUserMessageRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CLup.API.Contracts.Messages.SendUserMessage.SendUserMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Message"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CLup.API.Contracts.Messages.MarkMessageAsDeletedForUser.MarkMessageAsDeletedForUserRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CLup.API.Contracts.Messages.MarkMessageAsDeletedForUser.MarkMessageAsDeletedForUserRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CLup.API.Contracts.Messages.MarkMessageAsDeletedForUser.MarkMessageAsDeletedForUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/message/business": {
      "patch": {
        "tags": [
          "Message"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CLup.API.Contracts.Messages.SendBusinessMessage.SendBusinessMessageRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CLup.API.Contracts.Messages.SendBusinessMessage.SendBusinessMessageRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CLup.API.Contracts.Messages.SendBusinessMessage.SendBusinessMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Message"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CLup.API.Contracts.Messages.MarkMessageAsDeletedForBusiness.MarkMessageAsDeletedForBusinessRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CLup.API.Contracts.Messages.MarkMessageAsDeletedForBusiness.MarkMessageAsDeletedForBusinessRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CLup.API.Contracts.Messages.MarkMessageAsDeletedForBusiness.MarkMessageAsDeletedForBusinessRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CLup.API.Contracts.Businesses.CreateBusiness.CreateBusinessRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "zip": {
            "type": "integer",
            "format": "int32"
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "longitude": {
            "type": "number",
            "format": "double"
          },
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "opens": {
            "$ref": "#/components/schemas/System.TimeOnly"
          },
          "closes": {
            "$ref": "#/components/schemas/System.TimeOnly"
          },
          "capacity": {
            "type": "integer",
            "format": "int32"
          },
          "timeSlotLengthInMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "$ref": "#/components/schemas/CLup.Domain.Businesses.Enums.BusinessType"
          }
        },
        "additionalProperties": false
      },
...
      "CLup.API.Contracts.Businesses.UpdateBusiness.UpdateBusinessRequest": {
        "type": "object",
        "properties": {
          "businessId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "zip": {
            "type": "integer",
            "format": "int32"
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "longitude": {
            "type": "number",
            "format": "double"
          },
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "opens": {
            "$ref": "#/components/schemas/System.TimeOnly"
          },
          "closes": {
            "$ref": "#/components/schemas/System.TimeOnly"
          },
          "capacity": {
            "type": "integer",
            "format": "int32"
          },
          "timeSlotLengthInMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "$ref": "#/components/schemas/CLup.Domain.Businesses.Enums.BusinessType"
          }
        },
        "additionalProperties": false
      },
...
      "CLup.API.Contracts.Messages.MarkMessageAsDeletedForBusiness.MarkMessageAsDeletedForBusinessRequest": {
        "type": "object",
        "properties": {
          "senderId": {
            "type": "string",
            "format": "uuid"
          },
          "messageId": {
            "type": "string",
            "format": "uuid"
          },
          "receivedMessage": {
            "type": "boolean",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CLup.API.Contracts.Messages.MarkMessageAsDeletedForUser.MarkMessageAsDeletedForUserRequest": {
        "type": "object",
        "properties": {
          "messageId": {
            "type": "string",
            "format": "uuid"
          },
          "receivedMessage": {
            "type": "boolean",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CLup.API.Contracts.Messages.SendBusinessMessage.SendBusinessMessageRequest": {
        "type": "object",
        "properties": {
          "senderId": {
            "type": "string",
            "format": "uuid"
          },
          "receiverId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/CLup.Domain.Messages.Enums.MessageType"
          }
        },
        "additionalProperties": false
      },
      "CLup.API.Contracts.Messages.SendUserMessage.SendUserMessageRequest": {
        "type": "object",
        "properties": {
          "receiverId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/CLup.Domain.Messages.Enums.MessageType"
          }
        },
        "additionalProperties": false
      },
...
      "CLup.Domain.Businesses.Enums.BusinessType": {
        "enum": [
          1,
          2,
          3,
          4,
          5,
          6,
          7
        ],
        "type": "integer",
        "format": "int32"
      },
      "CLup.Domain.Messages.Enums.MessageType": {
        "enum": [
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "Microsoft.AspNetCore.Mvc.ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "System.TimeOnly": {
        "type": "object",
        "properties": {
          "hour": {
            "type": "integer",
            "format": "int32"
          },
          "minute": {
            "type": "integer",
            "format": "int32"
          },
          "second": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "millisecond": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "microsecond": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "nanosecond": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "ticks": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}